# contracts/openapi/web.openapi.yaml
openapi: 3.1.0
info:
  title: Reup Dubbing Studio Control Plane API
  version: 0.1.0
  description: >-
    Contract JIT cho Web Dashboard và Control Plane. Các slice Health, Settings
    và Channel/Series Profile đã có provider integration, generated client và
    consumer test tương ứng.
servers:
  - url: http://localhost:3000/v1
    description: Môi trường phát triển cục bộ
paths:
...
  /videos/deletions:
    post:
      summary: Yêu cầu xóa vĩnh viễn nhiều video
      operationId: deleteVideos
      parameters: [{ $ref: '#/components/parameters/IngestIdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/VideoBulkDeletionRequest' } } } }
      responses:
        '200': { description: 'Kết quả theo từng video, đúng thứ tự request', content: { application/json: { schema: { $ref: '#/components/schemas/VideoBulkDeletionEnvelope' } } } }
        default: { $ref: '#/components/responses/Problem' }
  /videos/{videoId}:
    get:
      summary: Đọc Video Library
      operationId: getVideo
      parameters: [{ $ref: '#/components/parameters/VideoId' }]
      responses:
        '200': { description: Video detail, headers: { ETag: { $ref: '#/components/headers/VersionEtag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/VideoEnvelope' } } } }
        default: { $ref: '#/components/responses/Problem' }
    delete:
      summary: Yêu cầu xóa vĩnh viễn video; object và dữ liệu được dọn ở nền
      operationId: deleteVideo
      parameters:
        - { $ref: '#/components/parameters/VideoId' }
        - { $ref: '#/components/parameters/VideoIfMatch' }
        - { $ref: '#/components/parameters/IngestIdempotencyKey' }
      responses:
        '202': { description: Video đang được xóa ở nền, content: { application/json: { schema: { $ref: '#/components/schemas/VideoDeletionEnvelope' } } } }
        default: { $ref: '#/components/responses/Problem' }
...
  /videos/{videoId}/studio:
    get:
      summary: Đọc dữ liệu Studio
      operationId: getStudio
      parameters: [{ $ref: '#/components/parameters/VideoId' }]
      responses:
        '200': { description: Studio detail, headers: { ETag: { $ref: '#/components/headers/VersionEtag' } }, content: { application/json: { schema: { $ref: '#/components/schemas/StudioEnvelope' } } } }
        default: { $ref: '#/components/responses/Problem' }
  /videos/{videoId}/segments/{segmentId}:
    patch:
      summary: Tạo revision mới cho segment
      operationId: updateStudioSegment
      parameters: [{ $ref: '#/components/parameters/VideoId' }, { name: segmentId, in: path, required: true, schema: { $ref: '#/components/schemas/UuidV7' } }, { $ref: '#/components/parameters/IfMatch' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/StudioSegmentPatch' } } } }
      responses:
        '200': { description: Segment revision, content: { application/json: { schema: { $ref: '#/components/schemas/StudioMutationEnvelope' } } } }
        default: { $ref: '#/components/responses/Problem' }
  /videos/{videoId}/segments/{segmentId}/preview:
    post:
      summary: Xin preview audio của segment
      operationId: previewStudioSegment
      parameters: [{ $ref: '#/components/parameters/VideoId' }, { name: segmentId, in: path, required: true, schema: { $ref: '#/components/schemas/UuidV7' } }]
      responses:
        '200': { description: Preview metadata, content: { application/json: { schema: { $ref: '#/components/schemas/StudioAssetEnvelope' } } } }
        default: { $ref: '#/components/responses/Problem' }
  /videos/{videoId}/segments/{segmentId}/regenerate:
    post:
      summary: Tạo request regenerate segment
      operationId: regenerateStudioSegment
      parameters: [{ $ref: '#/components/parameters/VideoId' }, { name: segmentId, in: path, required: true, schema: { $ref: '#/components/schemas/UuidV7' } }, { $ref: '#/components/parameters/IfMatch' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      responses:
        '202': { description: Regenerate queued, content: { application/json: { schema: { $ref: '#/components/schemas/StudioMutationEnvelope' } } } }
        default: { $ref: '#/components/responses/Problem' }
  /videos/{videoId}/review-decisions:
    post:
      summary: Lưu review decision
      operationId: createStudioReviewDecision
      parameters: [{ $ref: '#/components/parameters/VideoId' }, { $ref: '#/components/parameters/IfMatch' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/StudioReviewRequest' } } } }
      responses:
        '200': { description: Review decision, content: { application/json: { schema: { $ref: '#/components/schemas/StudioMutationEnvelope' } } } }
        default: { $ref: '#/components/responses/Problem' }
  /videos/{videoId}/render-requests:
    post:
      summary: Yêu cầu render output
      operationId: requestStudioRender
      parameters: [{ $ref: '#/components/parameters/VideoId' }, { $ref: '#/components/parameters/IfMatch' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      responses:
        '202': { description: Render queued, content: { application/json: { schema: { $ref: '#/components/schemas/StudioMutationEnvelope' } } } }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  parameters:
    VideoId:
      name: videoId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/UuidV7' }
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag nhận từ lần đọc Settings gần nhất.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
    IngestIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128, pattern: '^[!-~]+$' }
...
    VideoIfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag nhận từ lần đọc video gần nhất.
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
...
  headers:
    VersionEtag:
      description: Strong ETag của aggregate version.
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
    RequestId:
      description: UUID v7 do API sinh cho request hiện tại.
      schema:
        $ref: '#/components/schemas/RequestId'
...
  schemas:
...
    VideoOutputReadiness:
      type: string
      enum: [NONE, PARTIAL, READY, CLEANED]
...
    VideoEnvelope:
      type: object
      required: [data]
      properties: { data: { $ref: '#/components/schemas/Video' } }
    VideoDeletion:
      type: object
      additionalProperties: false
      required: [requestedAt, errorCode]
      properties:
        requestedAt: { type: string, format: date-time }
        errorCode: { type: [string, 'null'] }
    VideoDeletionResult:
      type: object
      additionalProperties: false
      required: [videoId, status, cancelledJobIds]
      properties:
        videoId: { $ref: '#/components/schemas/UuidV7' }
        status: { type: string, enum: [DELETING, DELETE_FAILED] }
        cancelledJobIds: { type: array, items: { $ref: '#/components/schemas/UuidV7' } }
    VideoDeletionEnvelope:
      type: object
      required: [data]
      properties: { data: { $ref: '#/components/schemas/VideoDeletionResult' } }
    VideoBulkDeletionRequest:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            additionalProperties: false
            required: [videoId, version]
            properties:
              videoId: { $ref: '#/components/schemas/UuidV7' }
              version: { type: integer, minimum: 1 }
    VideoBulkDeletionItemResult:
      type: object
      additionalProperties: false
      required: [videoId, result, cancelledJobIds]
      properties:
        videoId: { $ref: '#/components/schemas/UuidV7' }
        result: { type: string, enum: [ACCEPTED, ALREADY_DELETING, HAS_PUBLICATION_HISTORY, VERSION_CONFLICT, NOT_FOUND] }
        cancelledJobIds: { type: array, items: { $ref: '#/components/schemas/UuidV7' } }
    VideoBulkDeletionResult:
      type: object
      additionalProperties: false
      required: [items]
      properties: { items: { type: array, items: { $ref: '#/components/schemas/VideoBulkDeletionItemResult' } } }
    VideoBulkDeletionEnvelope:
      type: object
      required: [data]
      properties: { data: { $ref: '#/components/schemas/VideoBulkDeletionResult' } }
    VideoOutputPart:
      description: Asset metadata of one render output part. Fetch the file itself with grantVideoOutput.
      type: object
      additionalProperties: false
      required: [assetId, status, contentType, fileName]
      properties:
        assetId: { $ref: '#/components/schemas/UuidV7' }
        status: { type: string }
        contentType: { type: [string, 'null'] }
        fileName: { type: [string, 'null'] }
    VideoOutput:
      type: object
      additionalProperties: false
      required: [id, variant, revision, status, video, subtitle, thumbnail]
      properties:
        id: { $ref: '#/components/schemas/UuidV7' }
        variant: { type: string }
        revision: { type: integer }
        status: { type: string }
        video: { oneOf: [{ $ref: '#/components/schemas/VideoOutputPart' }, { type: 'null' }] }
        subtitle: { oneOf: [{ $ref: '#/components/schemas/VideoOutputPart' }, { type: 'null' }] }
        thumbnail: { oneOf: [{ $ref: '#/components/schemas/VideoOutputPart' }, { type: 'null' }] }
    Video:
      type: object
      required: [id, version, displayTitle, status, sourceLanguage, targetLanguage, source, profile, latestJob, reviewStatus, outputSummary, thumbnail, outputs, assets, createdAt, updatedAt, ingestedAt, archivedAt, deletion, capabilities]
      properties:
        id: { $ref: '#/components/schemas/UuidV7' }
        version: { type: integer }
        displayTitle: { type: [string, 'null'] }
        status: { type: string }
        sourceLanguage: { type: string }
        targetLanguage: { type: string }
        source:
          type: object
          required: [platform, externalId, canonicalUrl, durationMs, creatorName]
          properties: { platform: { type: string }, externalId: { type: string }, canonicalUrl: { type: [string, 'null'] }, durationMs: { type: [integer, 'null'] }, creatorName: { type: [string, 'null'] } }
        profile:
          type: object
          required: [channelProfileId, channelProfileName, seriesProfileId, seriesProfileName]
          properties: { channelProfileId: { $ref: '#/components/schemas/UuidV7' }, channelProfileName: { type: string }, seriesProfileId: { oneOf: [{ $ref: '#/components/schemas/UuidV7' }, { type: 'null' }] }, seriesProfileName: { type: [string, 'null'] } }
        latestJob:
          oneOf:
            - type: object
              required: [id, kind, status, progress, currentTask, failure, updatedAt]
              properties: { id: { $ref: '#/components/schemas/UuidV7' }, kind: { type: string }, status: { type: string }, progress: { type: integer }, currentTask: { type: [string, 'null'] }, failure: { type: [object, 'null'], properties: { code: { type: string }, detail: { type: [string, 'null'] } }, required: [code, detail] }, updatedAt: { type: string, format: date-time } }
            - { type: 'null' }
        reviewStatus: { type: string, enum: [NOT_REQUIRED, PENDING, CHANGES_REQUESTED, APPROVED] }
        outputSummary: { type: object, required: [readiness, requiredVariants, availableVariants, warnings], properties: { readiness: { $ref: '#/components/schemas/VideoOutputReadiness' }, requiredVariants: { type: array, items: { type: string } }, availableVariants: { type: array, items: { type: string } }, warnings: { type: array, items: { type: string } } } }
        thumbnail: { type: [object, 'null'], additionalProperties: true }
        outputs: { type: array, items: { $ref: '#/components/schemas/VideoOutput' } }
        assets: { type: array, items: { type: object, additionalProperties: true } }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        ingestedAt: { type: [string, 'null'], format: date-time }
        archivedAt: { type: [string, 'null'], format: date-time }
        deletion: { oneOf: [{ $ref: '#/components/schemas/VideoDeletion' }, { type: 'null' }] }
        capabilities:
          type: object
          required: [canOpenStudio, canOpenPublishing, canArchive, canDelete, deleteBlockedReason]
          properties:
            canOpenStudio: { type: boolean }
            canOpenPublishing: { type: boolean }
            canArchive: { type: boolean }
            canDelete: { type: boolean }
            deleteBlockedReason: { type: [string, 'null'], enum: [PUBLICATION_HISTORY, DELETING, null] }
    StudioSegmentPatch:
      type: object
      minProperties: 1
      properties: { translatedText: { type: string, maxLength: 5000 }, castSheetEntryId: { oneOf: [{ $ref: '#/components/schemas/UuidV7' }, { type: 'null' }] }, voiceProfileId: { $ref: '#/components/schemas/UuidV7' }, targetStartMs: { type: integer, minimum: 0 }, targetEndMs: { type: integer, minimum: 1 }, speechRate: { type: number, minimum: 0.5, maximum: 2 }, editReason: { type: string, maxLength: 500 } }
    StudioReviewRequest:
      type: object
      required: [scope, subjectVersion, decision]
      properties: { scope: { type: string, enum: [CAST, SCRIPT, TTS, RENDER] }, subjectVersion: { type: string }, decision: { type: string, enum: [APPROVED, CHANGES_REQUESTED, REJECTED] }, note: { type: string, maxLength: 2000 } }
    StudioMutationEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [version]
          properties:
            videoId: { $ref: '#/components/schemas/UuidV7' }
            segmentId: { $ref: '#/components/schemas/UuidV7' }
            revisionId: { $ref: '#/components/schemas/UuidV7' }
            revision: { type: integer, minimum: 1 }
            jobId: { $ref: '#/components/schemas/UuidV7' }
            taskId: { $ref: '#/components/schemas/UuidV7' }
            id: { $ref: '#/components/schemas/UuidV7' }
            scope: { type: string }
            subjectVersion: { type: string }
            decision: { type: string }
            note: { type: [string, 'null'] }
            decidedAt: { type: string, format: date-time }
            version: { type: integer, minimum: 1 }
    StudioAssetEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [method, url, expiresAt, fileName, contentType, byteSize]
          properties:
            method: { type: string, enum: [GET] }
            url: { type: string, format: uri }
            expiresAt: { type: string, format: date-time }
            fileName: { type: string }
            contentType: { type: [string, 'null'] }
            byteSize: { type: [string, 'null'] }
    StudioEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [video, transcriptRuns, cast, segments, reviews, capabilities]
          properties:
            video: { $ref: '#/components/schemas/StudioVideo' }
            transcriptRuns: { type: array, items: { $ref: '#/components/schemas/StudioTranscriptRun' } }
            cast: { oneOf: [{ $ref: '#/components/schemas/StudioCastSheet' }, { type: 'null' }] }
            segments: { type: array, items: { $ref: '#/components/schemas/StudioSegment' } }
            reviews: { type: array, items: { $ref: '#/components/schemas/StudioReviewDecision' } }
            capabilities: { $ref: '#/components/schemas/StudioCapabilities' }
    StudioVideo:
      type: object
      required: [id, version, title, status, sourceDurationMs]
      properties:
        id: { $ref: '#/components/schemas/UuidV7' }
        version: { type: integer, minimum: 1 }
        title: { type: [string, 'null'] }
        status: { type: string }
        sourceDurationMs: { type: [integer, 'null'], minimum: 1 }
    StudioTranscriptRun:
      type: object
      required: [id, method, status, language, averageConfidence, selectedAt]
      properties:
        id: { $ref: '#/components/schemas/UuidV7' }
        method: { type: string }
        status: { type: string }
        language: { type: string }
        averageConfidence: { type: [string, 'null'] }
        selectedAt: { type: [string, 'null'], format: date-time }
    StudioCastSheet:
      type: object
      required: [id, status, version, entries]
      properties:
        id: { $ref: '#/components/schemas/UuidV7' }
        status: { type: string }
        version: { type: integer, minimum: 1 }
        entries: { type: array, items: { $ref: '#/components/schemas/StudioCastEntry' } }
    StudioCastEntry:
      type: object
      required: [id, characterKey, displayName, roleKind, voice]
      properties:
        id: { $ref: '#/components/schemas/UuidV7' }
        characterKey: { type: string }
        displayName: { type: string }
        roleKind: { type: string }
        voice:
          type: object
          required: [id, name]
          properties: { id: { $ref: '#/components/schemas/UuidV7' }, name: { type: string } }
    StudioSegment:
      type: object
      required: [id, ordinal, sourceStartMs, sourceEndMs, revision]
      properties:
        id: { $ref: '#/components/schemas/UuidV7' }
        ordinal: { type: integer, minimum: 1 }
        sourceStartMs: { type: integer, minimum: 0 }
        sourceEndMs: { type: integer, minimum: 1 }
        revision: { oneOf: [{ $ref: '#/components/schemas/StudioSegmentRevision' }, { type: 'null' }] }
    StudioSegmentRevision:
      type: object
      required: [id, revision, sourceText, translatedText, targetStartMs, targetEndMs, speechRate, status, castSheetEntryId, voice, preview]
      properties:
        id: { $ref: '#/components/schemas/UuidV7' }
        revision: { type: integer, minimum: 1 }
        sourceText: { type: string }
        translatedText: { type: string }
        targetStartMs: { type: integer, minimum: 0 }
        targetEndMs: { type: integer, minimum: 1 }
        speechRate: { type: number, minimum: 0.5, maximum: 2 }
        status: { type: string }
        castSheetEntryId: { oneOf: [{ $ref: '#/components/schemas/UuidV7' }, { type: 'null' }] }
        voice:
          type: object
          required: [id, name]
          properties: { id: { $ref: '#/components/schemas/UuidV7' }, name: { type: string } }
        preview: { oneOf: [{ $ref: '#/components/schemas/StudioPreview' }, { type: 'null' }] }
    StudioPreview:
      type: object
      required: [id, status, durationMs]
      properties: { id: { $ref: '#/components/schemas/UuidV7' }, status: { type: string }, durationMs: { type: [integer, 'null'] } }
    StudioReviewDecision:
      type: object
      required: [id, scope, subjectVersion, decision, note, decidedAt]
      properties: { id: { $ref: '#/components/schemas/UuidV7' }, scope: { type: string }, subjectVersion: { type: string }, decision: { type: string }, note: { type: [string, 'null'] }, decidedAt: { type: string, format: date-time } }
    StudioCapabilities:
      type: object
      required: [canEdit, canRegenerate, canRender]
      properties: { canEdit: { type: boolean }, canRegenerate: { type: boolean }, canRender: { type: boolean } }
...
    UuidV7:
      type: string
      format: uuid
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
...
    RequestId:
      type: string
      format: uuid
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
...
    ProblemDetails:
      type: object
      description: RFC 9457 với extension code và requestId.
      required:
        - type
        - title
        - status
        - instance
        - code
        - requestId
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - ROUTE_NOT_FOUND
            - RATE_LIMITED
            - INTERNAL_ERROR
            - SETTINGS_NOT_CONFIGURED
            - SETTINGS_VALIDATION_FAILED
            - CONNECTION_TEST_FAILED
            - VERSION_CONFLICT
            - PROFILE_NAME_CONFLICT
            - PROFILE_NOT_FOUND
            - PROFILE_ARCHIVED
            - PROFILE_NOT_READY
            - PROFILE_HAS_ACTIVE_SERIES
            - PROFILE_PARENT_ARCHIVED
            - PROFILE_VERSION_CONFLICT
            - PROFILE_ASSET_NOT_AVAILABLE
            - PROFILE_ASSET_ROLE_INVALID
            - PROFILE_MASK_INVALID
            - PROFILE_VOICE_NOT_READY
            - PROFILE_VALIDATION_FAILED
            - VOICE_NOT_FOUND
            - VOICE_NAME_CONFLICT
            - VOICE_VERSION_CONFLICT
            - VOICE_VALIDATION_FAILED
            - VOICE_NOT_READY
            - VOICE_ARCHIVED
            - VOICE_IN_USE
            - VOICE_SAMPLE_NOT_AVAILABLE
            - SOURCE_ACCOUNT_NOT_FOUND
            - SOURCE_ACCOUNT_VERSION_CONFLICT
            - SOURCE_ACCOUNT_CREDENTIAL_REQUIRED
            - DISCOVERY_RUN_NOT_FOUND
            - DISCOVERY_MODE_DISABLED
            - DISCOVERY_RUN_CONFLICT
            - WATCHLIST_NOT_FOUND
            - WATCHLIST_VERSION_CONFLICT
            - WATCHLIST_DUPLICATE
            - WATCHLIST_RUN_ACTIVE
            - DISCOVERY_VALIDATION_FAILED
            - DISCOVERY_CURSOR_INVALID
            - DISCOVERY_PROVIDER_UNAVAILABLE
            - INGEST_VALIDATION_FAILED
            - INGEST_NO_CREATABLE_ITEMS
            - QUEUE_VALIDATION_FAILED
            - QUEUE_JOB_NOT_FOUND
            - QUEUE_CURSOR_INVALID
            - JOB_NOT_CANCELLABLE
            - JOB_NOT_RETRYABLE
            - JOB_RETRY_CONFLICT
            - IDEMPOTENCY_KEY_REUSED
            - WORKER_VALIDATION_FAILED
            - WORKER_NOT_FOUND
            - WORKER_NOT_ACTIVE
            - WORKER_NOT_DRAINABLE
            - WORKER_NOT_SAFE_TO_TERMINATE
            - WORKER_SESSION_CONFLICT
            - WORKER_VERSION_MISMATCH
            - WORKER_IMAGE_NOT_APPROVED
            - WORKER_HARDWARE_MISMATCH
            - ENROLLMENT_TOKEN_INVALID
            - ENROLLMENT_TOKEN_EXPIRED
            - ENROLLMENT_TOKEN_CONSUMED
            - WORKER_CREDENTIAL_REVOKED
            - PUBLISH_PACKAGE_NOT_FOUND
            - PUBLICATION_TASK_NOT_FOUND
            - PUBLICATION_VERSION_CONFLICT
            - PUBLICATION_INVALID_TRANSITION
            - PUBLICATION_CONTENT_INCOMPLETE
            - PUBLICATION_CONTENT_APPROVAL_REQUIRED
            - PUBLICATION_CHECKLIST_INCOMPLETE
            - PUBLICATION_PROOF_REQUIRED
            - PUBLICATION_ASSET_UNAVAILABLE
            - PUBLICATION_DESTINATION_INACTIVE
            - PUBLICATION_FIELD_LOCKED
            - PUBLICATION_CONTENT_AGENT_UNAVAILABLE
            - VIDEO_NOT_FOUND
            - VIDEO_VERSION_CONFLICT
            - VIDEO_HAS_PUBLICATION_HISTORY
            - VIDEO_DELETING
            - VIDEO_DELETION_VALIDATION_FAILED
        requestId:
          $ref: '#/components/schemas/RequestId'
  responses:
    Problem:
      description: Lỗi HTTP chuẩn RFC 9457; không dùng success envelope.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'